+++ docs/changelog.rst
@@ -16,9 +16,7 @@ The third digit is only for regressions.
 Changes:
 ^^^^^^^^
 
-- Add optional_ validator.
-
-.. _optional: http://attrs.readthedocs.org/en/stable/api.html#validators
+- Add :func:`attr.validators.optional` that wraps other validators allowing attributes to be ``None``.
 
 
 15.0.0 (2015-04-15)